fix: fix ipc

+++ packages/kaitian-extension/src/node/extension.service.ts
@@ -5,7 +5,7 @@ import * as fs from 'fs-extra';
 import { Injectable, Autowired } from '@ali/common-di';
 import { ExtensionScanner } from './extension.scanner';
 import { IExtensionMetaData, IExtensionNodeService, ExtraMetaData } from '../common';
-import { getLogger, Deferred, isDevelopment, INodeLogger } from '@ali/ide-core-node';
+import { getLogger, Deferred, isDevelopment, INodeLogger, isWindows } from '@ali/ide-core-node';
 import * as cp from 'child_process';
 import * as psTree from 'ps-tree';
 import * as isRunning from 'is-running';
@@ -336,7 +336,9 @@ export class ExtensionNodeServiceImpl implements IExtensionNodeService  {
       console.log('mainThreadListenPath', mainThreadListenPath);
 
       try {
-        await fs.unlink(mainThreadListenPath);
+        if (!isWindows) {
+          await fs.unlink(mainThreadListenPath);
+        }
       } catch (e) {
         getLogger().error(e);
       }
@@ -428,7 +430,9 @@ export class ExtensionNodeServiceImpl implements IExtensionNodeService  {
       this.electronMainThreadServer = server;
       const listenPath = this.getElectronMainThreadListenPath(clientId);
       try {
-        await fs.unlink(listenPath);
+        if (!isWindows) {
+          await fs.unlink(listenPath);
+        }
       } catch (e) {
         getLogger().error(e);
       }
@@ -506,7 +510,9 @@ export class ExtensionNodeServiceImpl implements IExtensionNodeService  {
     const extServer = net.createServer();
 
     try {
-      await fs.unlink(extServerListenPath);
+      if (!isWindows) {
+        await fs.unlink(extServerListenPath);
+      }
     } catch (e) {}
 
     const extConnection =  await new Promise((resolve) => {
@@ -538,7 +544,9 @@ export class ExtensionNodeServiceImpl implements IExtensionNodeService  {
     this.clientExtProcessExtConnectionServer.set(clientId, extServer);
 
     try {
-      await fs.unlink(extServerListenPath);
+      if (!isWindows) {
+        await fs.unlink(extServerListenPath);
+      }
     } catch (e) { }
 
     const extConnection =  await new Promise((resolve) => {
